Usage Report Statistics
A lot of scripts, functions and aliases are not used. In the effort of
detecting which ones are used often and which ones not at all, a usage
statistics mechanism has been implemented.

The mechanism uses `preexec` built-in funcion of the `zsh` shell. This
funcion captures every command and filters down only custom:
- scripts
- functions
- aliases

The generated report is used to compare all the scripts, functions,
aliases and see which ones are used most often.