Custom Projects Completion
Adding abililty to use TAG in comments to create relationships between
various files that belong to the same `project`. List of projects comes
from a custom completion source and TAG is highlighted as part of the
todo plugin.

This creates well integrated experience with the ability to use
Telescope and other builtin neovim features to manage related projects
as tags in otherwise decentralized files.