autokey: make web browser script keyboard friendly
Autokey script outputs a zenity dialog that allows for selecting
multiple choice actions to be performed on a web page (create link,
task, add to paylist or capture web highlight). The AutoKey python
wrapper does not work with keyboard shortcuts. This means that it's not
possible to select options from the dialog box with keys.

This change wraps zenity in a bash script which is required for adding
the ability to select options using keyboard. Now it's possible to
select each action by hitting first letter of the action:
- L: ink
- T: ask
- P: laylist
- H: ighlight

The ability to select multiple actions remainst the same but needs mouse
and Ctrl + click to select more than one action.

Signed-off-by: Piotr <decoder@live.de>