git: project git-log
Using logg function, currently mapped to Ctrl+l, it is now possible to:
- show all commits with a preview window in terminal
- search commit messages, SHA using fzf
- on-enter see commit details in a custom nvim buffer
- on ctrl+e open commit url in a default browser

The implementation adds a new git alias for formatting logs, custom
function and zsh key binding.

This improves commit search and navigation, which helps in inspecting a
git repo and reasoning about a code-base or lerning a new one.

Signed-off-by: Piotr <decoder@live.de>